Move `Weak` into a separate module
